remove unused code and simplify tests
